moving pipeline traits/classes to their own file
